+++ Comm2Tender/Logic/Models/Dto/Tender.cs
@@ -2,7 +2,7 @@
 {
     public class Tender
     {
-        public long TenderId { get; set; } // int
+        public int TenderId { get; set; } // int
         public string Number { get; set; } // nvarchar(50)
         public string Discription { get; set; } // nvarchar(50)
         public PercentsDictionary PercentsDictionary { get; set; } // int
